requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
ffmpeg-python==0.2.0
srt==3.5.2
pydantic==1.10.13
requests==2.31.0 